% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{calc_pest_phenology}
\alias{calc_pest_phenology}
\title{Calculate Pest Phenology}
\usage{
calc_pest_phenology(
  trap_data,
  pest = NULL,
  lat = NULL,
  lon = NULL,
  weather_source = "open_meteo",
  cimis_csv_path = NULL,
  custom_lower = NULL,
  custom_upper = NULL
)
}
\arguments{
\item{trap_data}{A data frame containing trap counts}

\item{pest}{Name of the pest}

\item{lat}{Latitude}

\item{lon}{Longitude}

\item{weather_source}{Weather data source}

\item{cimis_csv_path}{Path to the CIMIS csv file}

\item{custom_lower}{Lower developmental threshold}

\item{custom_upper}{Upper developmental threshold}
}
\description{
Calculate Pest Phenology
}
