% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{fetch_open_meteo}
\alias{fetch_open_meteo}
\title{Fetch Open-Meteo Data}
\usage{
fetch_open_meteo(lat, lon, start_date, end_date = as.character(Sys.Date()))
}
\arguments{
\item{lat}{Latitude}

\item{lon}{Longitude}

\item{start_date}{Start date format (YYYY-MM-DD)}

\item{end_date}{End date format (YYYY-MM-DD)}
}
\value{
A dataframe with Date, tmax (F), and tmin (F)
}
\description{
Retrieves weather data from the Open-Meteo API.
}
