% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pest_thresholds}
\alias{pest_thresholds}
\title{Pest Developmental Thresholds}
\format{
An object of class \code{data.frame} with 104 rows and 7 columns.
}
\usage{
pest_thresholds
}
\description{
A dataset containing the lower and upper developmental temperature
thresholds (in Fahrenheit) for common agricultural pests in California.
}
\details{
\describe{
\item{pest_code}{Short abbreviation for the pest (e.g., OLFF, NOW)}
\item{pest_name}{Common name of the pest}
\item{lower_thresh}{Lower developmental temperature threshold (°F)}
\item{upper_thresh}{Upper developmental temperature threshold (°F)}
}
}
\keyword{datasets}
