% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{plot_trap_phenology}
\alias{plot_trap_phenology}
\title{Plot Trap Phenology with Flight Markers}
\usage{
plot_trap_phenology(pheno_data, pest = NULL)
}
\arguments{
\item{pheno_data}{Phenology data frame}

\item{pest}{Name of the pest}
}
\description{
Plot Trap Phenology with Flight Markers
}
